Add initial support for mouse events and update example.
We now support mouseX, mouseY, pmouseX, pmouseY, mouseMove,
mouseDragged, mouseUp and mouseDown.
We store the global mouse positions in env for now, and pass it to
mouseX (for example) which is implemented as a simple function that
reads from env.

The example was updated to be a drag and drop one.

This won't work off the bat as the sdl y axis is flipped. We need a way
to tell sdl to flip the axis.