Update link flags to link against the C part of reglnative.
This looks like a tiny change on reprocessing, which is awesome. But it
actually hides a big change in reglnative.
We added a C dependency in reglnative called SOIL. This library allows
us to load a whole bunch of different file formats. We wrote a wrapper
that will basically return an `array int` representing the colors of
the image. It also returns the width, height and the channels (1, 2, 3
or 4) that the image has.

I think Rebel should be able to pick this dependency up and add it.
The way to do it:
- skim through the explicit dependencies
- for each check if they have a `dependencyName.pc` file
- make sure they have a `PKG_CONFIG_PATH` in their package.json
- add to the final link command `-cclib \"$(pkg-config --libs
dependencyName)\"`

Another way would be to actually run `pkg-config` ahead of time, which
would allow us to cache it. We'd also want rebel to cache for each dep
the output of that command. So the first build's a bit slow but we make
up for it later.