Add Image support. Fix alpha handling. Enable blending.
This is a big change. It adds supports for GL texture, which you can
load with `loadImage`. That returns a ref which will get magically
updated when the image's ready (loaded async in web but sync on native).
That ref can be passed to `image img 0 0` which will wait until the
image's ready and start rendering it.
We support alpha/blending. And we support loading BMP, PNG, JPG thanks
to SOIL for native.