Optimization pass.
This is a pretty simple optimization pass.
Instead of sending 3 different linear arrays in different vertex buffers
we make one single huge array and use the `stride` and `offset` values
that you can pass to `vertexAttribPointer`.

`offset` is the number of bytes from 0 from which the data starts for
the given attribute
`stride` is the number of bytes that you need to jump by to get the
next piece of data for the same kind
`size` is the number of data points (which isn't in byte but in units
independent from `type_`).

Here's a great graphical explanation:
https://learnopengl.com/img/getting-started/vertex_attribute_pointer_interleaved_textures.png

We pack x1, y1, z1, r1, g1, b1, a1, s1, t1 for each point. They're each
floats for we multiply by 4.
(s and t are the 2D coordinates used to map the texture onto the
geometry, numbers between 0 and 1).

I also inlined the draw code for drawing the image (instead of calling a
general purpose drawRectInternal). That removes one conditional / branch
from the draw code.

A further optimization would be to eliminate any allocation, and reuse
the same array.