Support loading multiple images. Update color -> colorT.
We didn't support rendering multiple images because we'd use the same
textureBuffer and didn't copy the data over for each image. Instead of
having one texture buffer and copy the data at every frame, we allocate
a new buffer for each image and copy the data once. Then it's just a
matter of binding it. We don't technically need the image data itself,
as it's just duplicated data. We could store width/height and texture
buffer instead of Gl.imageT. This was easier for now.