Implement batching for everything.
We flush the geometry before rendering textures / images (and so text).

We're using 2 global arrays, one for vertex data and one for elements
(indices of vertices). GL checks the element array to understand what
triangles we're trying to draw, and then uses that to lookup the actual
vertex to draw. This helps us send a bit less data at the end of each
render loop.

This has 1 big performance bottleneck in web: addRectToGlobalBatch.
The reason it's a performance issue is because `set` isn't inlined
correctly. I'm not sure why... It should be inlined. In my tests
with running bsc manually, I cannot make it _not_ inline the function
so not sure what's going on. Once we fix this, we get 3x perf
improvement.