Update playground + add script.
Updating this was a bit harder than expected.
There is code to be added at the start of reprocessing to ensure that
it's reentrant (it deletes the old canvas, adds a new one and makes
sure to give it a recognizable ID).
OCaml's pretty printer is also wrong for ## (or Reaosn has the wrong
AST).