Fix playground.
This commit will make our lives easier. All we need to do now is to run
the `eject.sh` script which bundles all of the library in a consumable
form (consumable by our playground). It does more than bundle though,
it'll do a simple string replace to make a call to ClientWrapper.init
instead of Gl.Window.init. That code's written in reason too.
The file generated needs to be pushed so that github pages can find it
but it'll be loaded on page load by the playground.