Add new createImage and withImage calls.
This is a new API that we're exposing to allow people to make draw calls
that are stored into an image which can then be drawn onto the screen.
This effectively allows you to drawn directly onto a loaded image, or
allows you to cache static draw operations.

cc @jaredly